Fixed Major Bug
Fixed a bug that would cause files to be deleted when using the folder
combiner. I was experimenting with removing the Create Folder function
and relying only on the Rename Folder function. Program should now work
for all cases: Folder is renamed, deleted, or created, and files inside
are renamed too.